Organise tests in feature folders
