--- src/__tests__/__fixtures__/pages/blog/99.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { PropsPrinter } from '../../../__utils__';
-
-export default function blog_99(props) {
-  return (
-    <>
-      /blog/99 -
-      <PropsPrinter props={props} />
-    </>
-  );
-}
