--- src/__tests__/__fixtures__/pages/blog/[id].js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { RouterQueryPrinter } from '../../../__utils__';
-
-export default function blog_$id$({ routerMock }) {
-  return (
-    <>
-      /blog/[id] -
-      <RouterQueryPrinter routerMock={routerMock} />
-    </>
-  );
-}
