--- src/__tests__/__fixtures__/pages/blog/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { PropsPrinter } from '../../../__utils__';
-
-export default function blog_index(props) {
-  return (
-    <>
-      /blog/index -
-      <PropsPrinter props={props} />
-    </>
-  );
-}
