--- src/__tests__/__fixtures__/pages/catch-all/[id]/[...slug]/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { RouterQueryPrinter } from '../../../../../__utils__/';
-
-export default function CatchAll_$id$_$slug$({ routerMock }) {
-  return (
-    <>
-      /catch-all/[id]/[...slug]/index -
-      <RouterQueryPrinter routerMock={routerMock} />
-    </>
-  );
-}
