--- src/__tests__/__fixtures__/pages/gip/[id].js
@@ -1,16 +1,0 @@
-import React from 'react';
-import { PropsPrinter, sleep } from '../../../__utils__';
-
-export default function gip_$id$(props) {
-  return (
-    <>
-      /gip/[id] -
-      <PropsPrinter props={props} />
-    </>
-  );
-}
-
-gip_$id$.getInitialProps = async (ctx) => {
-  await sleep(1);
-  return ctx;
-};
