--- src/__tests__/__fixtures__/pages/index.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import { PropsPrinter } from '../../__utils__';
-
-export default function index(props) {
-  return (
-    <>
-      /index -
-      <PropsPrinter props={props} />
-    </>
-  );
-}
