--- src/__tests__/__fixtures__/pages/ssr/redirect.js
@@ -1,15 +1,0 @@
-import { sleep } from '../../../__utils__/';
-
-export default function ssr_notFound() {
-  return `/ssr/redirect`;
-}
-
-export async function getServerSideProps(ctx) {
-  await sleep(1);
-  return {
-    redirect: {
-      destination: '/another-page',
-      permanent: false,
-    },
-  };
-}
