+++ src/__tests__/client-navigation/__fixtures__/pages/ssr/b.js
