+++ src/__tests__/page-data-fetching/__fixtures__/pages/ssr/[id].js
@@ -1,0 +1,22 @@
+import React from 'react';
+import { PropsPrinter, sleep } from '../../../../__utils__';
+
+export default function ssr_$id$(props) {
+  return (
+    <>
+      /ssr/[id]
+      <PropsPrinter props={props} />
+    </>
+  );
+}
+
+export async function getServerSideProps(ctx) {
+  await sleep(1);
+  return {
+    props: {
+      ...ctx,
+      isWindowDefined: global.window !== undefined,
+      isDocumentDefined: global.document !== undefined,
+    },
+  };
+}
