+++ src/__tests__/page-data-fetching/__fixtures__/pages/ssr/not-found.js
@@ -1,0 +1,12 @@
+import { sleep } from '../../../../__utils__';
+
+export default function ssr_notFound() {
+  return `/ssr/not-found`;
+}
+
+export async function getServerSideProps(ctx) {
+  await sleep(1);
+  return {
+    notFound: true,
+  };
+}
