+++ src/__tests__/page-data-fetching/__fixtures__/pages/ssr/redirect.js
@@ -1,0 +1,15 @@
+import { sleep } from '../../../../__utils__/';
+
+export default function ssr_notFound() {
+  return `/ssr/redirect`;
+}
+
+export async function getServerSideProps(ctx) {
+  await sleep(1);
+  return {
+    redirect: {
+      destination: '/another-page',
+      permanent: false,
+    },
+  };
+}
