+++ src/__tests__/router-mocking/__fixtures__/pages/with-router/[id].js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { RouterPrinter } from '../../../../__utils__';
+
+export default function WithRouter({ routerMock }) {
+  return <RouterPrinter routerMock={routerMock} />;
+}
