+++ src/__tests__/routing/__fixtures__/pages/blog/[id].js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { RouterQueryPrinter } from '../../../../__utils__';
+
+export default function blog_$id$({ routerMock }) {
+  return (
+    <>
+      /blog/[id] -
+      <RouterQueryPrinter routerMock={routerMock} />
+    </>
+  );
+}
