+++ src/__tests__/routing/__fixtures__/pages/blog/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { PropsPrinter } from '../../../../__utils__';
+
+export default function blog_index(props) {
+  return (
+    <>
+      /blog/index -
+      <PropsPrinter props={props} />
+    </>
+  );
+}
