+++ src/__tests__/routing/__fixtures__/pages/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import { PropsPrinter } from '../../../__utils__';
+
+export default function index(props) {
+  return (
+    <>
+      /index -
+      <PropsPrinter props={props} />
+    </>
+  );
+}
