+++ src/__tests__/routing/__fixtures__/pages/invalid-extension.foo
